refactor: make service exportable
